fixing main.py file
